b402/pyproject.toml
[tool.poetry]
name = "b402"
version = "0.1.0"
description = ""
authors = ["julio colomer <info@aiaccelera.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.4"
jupyterlab = "^4.2.4"
langchain = "^0.3.7"
langchain-core = "^0.3.20"
langchain-openai = "^0.2.0"
langchain-google-genai = "^2.1.12"
python-dotenv = "^1.0.1"
langchain-groq = "^0.3.8"
langchain-community = "^0.3.29"
pypdf = "^4.3.1"
wikipedia = "^1.4.0"
langchain-chroma = "^0.1.2"
faiss-cpu = "^1.8.0.post1"
bs4 = "^0.0.2"
unstructured = "^0.15.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
